1. Linked User and Product models through Eloquent relationships 2. Created factory and seeder content for users and products 3. API routes: Addition of restful routes for products 4. Addition of products controller 5. Update to products schema

+++ database/migrations/2018_07_13_183519_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             
             $table->string('title');
             $table->text('description');
-            $table->integer('price');
+            $table->float('price', 8, 2);
             $table->boolean('availability');
 
             $table->timestamps();
